velox_graph
Struct VeloxGraph

pub struct VeloxGraph<NodeT: Clone, ConnectionT: Clone> {
    pub num_entries: usize,
    /* private fields */
}

Fields
num_entries: usize
Implementations

impl<NodeT: Clone, ConnectionT: Clone> VeloxGraph<NodeT, ConnectionT>

pub fn new() -> VeloxGraph<NodeT, ConnectionT>

Initialize the graph.

Example

// INFO: Sample data to store in the nodes. This is CUSTOM DATA defined by you that is stored in each node.
#[derive(Clone, Debug)]
struct NodeData {
    x: u32,
    y: u32,
}

// INFO: Sample data to store in the connections. This is CUSTOM DATA defined by you that is stored in each connection.
#[derive(Clone, Debug)]
struct ConnData {
    a: u32,
    b: f64,
}

// INFO: Initialize the graph with types NodeData, for nodes, and ConnData, for connections.
let mut graph: VeloxGraph<NodeData, ConnData> = VeloxGraph::new();
assert_eq!(graph.num_entries, 0);

pub fn node_create(&mut self, node_data: NodeT) -> usize

Create nodes.

Example

// INFO: Initialize the graph.
let mut graph: VeloxGraph<u32, f64> = VeloxGraph::new();

// INFO: Create your first nodes.
let node_id0 = graph.node_create(634);
let node_id1 = graph.node_create(43);

assert_eq!(graph.num_entries, 2);

pub fn node_get<'a>( &'a mut self, node_id: usize, ) -> Result<&'a mut Node<NodeT, ConnectionT>, VeloxGraphError>

Get mutable access to a node.

Example

// INFO: Initialize the graph.
let mut graph: VeloxGraph<u32, f64> = VeloxGraph::new();

// INFO: Create a node.
let node_id = graph.node_create(4);

// INFO: Get a mutable reference to that node.
let node = graph.node_get(node_id).unwrap();

assert_eq!(node.data, 4);

// INFO: Make changes to the node.
node.data += 5

assert_eq!(node.data, 9);

pub fn node_delete( &mut self, node_id_to_delete: usize, ) -> Result<(), VeloxGraphError>

Delete nodes.

Example

// INFO: Initialize the graph.
let mut graph: VeloxGraph<u32, f64> = VeloxGraph::new();

// INFO: Create a node.
let node_id = graph.node_create(634);
assert_eq!(graph.num_entries, 1);

// INFO: Delete the node. Its connections are automatically deleted as well.
graph.node_delete(node_id).unwrap();

assert_eq!(graph.num_entries, 0);

pub fn nodes_connection_set( &mut self, first_node_id: usize, second_node_id: usize, connection_data: ConnectionT, ) -> Result<(), VeloxGraphError>

Create node connections.

Example

// INFO: Initialize the graph.
let mut graph: VeloxGraph<u32, f64> = VeloxGraph::new();

// INFO: Create example nodes.
let node_id0 = graph.node_create(634);
let node_id1 = graph.node_create(43);

// INFO: Create connection from node0 to node1.
graph.nodes_connection_set(node_id0, node_id1, 5.24).unwrap();

// INFO: Get a mutable reference to that node.
let node0 = graph.node_get(node_id0).unwrap();

assert_eq!(node0.connections_forward_get_all().data_vec.len(), 1);

pub fn nodes_connection_remove( &mut self, first_node_id: usize, second_node_id: usize, ) -> Result<(), VeloxGraphError>

Delete node connections.

Example

// INFO: Initialize the graph.
let mut graph: VeloxGraph<u32, f64> = VeloxGraph::new();

// INFO: Create example nodes.
let node_id0 = graph.node_create(634);
let node_id1 = graph.node_create(43);

// INFO: Create connection from node0 to node1.
graph.nodes_connection_set(node_id0, node_id1, 5.24).unwrap();

// INFO: Get a mutable reference to that node.
let node0 = graph.node_get(node_id0).unwrap();
assert_eq!(node0.connections_forward_get_all().data_vec.len(), 1);

// INFO: Delete node connection.
graph.nodes_connection_remove(node_id0, node_id1).unwrap();

assert_eq!(node0.connections_forward_get_all().data_vec.len(), 0);